requirements.txt
# oTree-may-not-overwrite-this-file
otree==5.10.4
psycopg2>=2.8.4
sentry-sdk>=0.7.9
Django>=3.1,<4.0
channels>=3.0
channels_redis>=3.2
google-auth-oauthlib==0.4.6
google-auth==1.35.0
google-api-python-client==2.38.0
httplib2==0.20.4
requests==2.27.1
daphne>=3.0
whitenoise>=5.2.0
dj-database-url==0.5.0
psycopg2-binary>=2.9.3
gunicorn>=20.1.0
redis>=4.0.2
Pillow>=8.3.2